Generate/validate help for Debug and Release, fix issue with error not being thrown during validation

+++ tools/GenerateHelp.ps1
@@ -11,7 +11,7 @@ Param(
 
 Import-Module "$PSScriptRoot\HelpGeneration\HelpGeneration.psm1"
 
-$HelpFolders = Get-ChildItem "help" -Recurse -Directory | where { $_.FullName -like "*Debug*" }
+$HelpFolders = Get-ChildItem "help" -Recurse -Directory | where { $_.FullName -like "*Debug*" -or $_.FullName -like "*Release*" }
 
 # ---------------------------------------------------------------------------------------------
 
@@ -19,13 +19,17 @@ if ($ValidateMarkdownHelp)
 {
     New-Item -Path "$PSScriptRoot\HelpGeneration\Exceptions" -Name NewValidateHelpExceptions.csv -ItemType File -Force | Out-Null
 
-    Add-Content "$PSscriptRoot\HelpGeneration\Exceptions\NewValidateHelpExceptions.csv" "Module,Target,Description"
+    Add-Content "$PSScriptRoot\HelpGeneration\Exceptions\NewValidateHelpExceptions.csv" "Module,Target,Description"
 
     $HelpFolders | foreach { Validate-MarkdownHelp $_ }
 
     $Exceptions = Import-Csv "$PSScriptRoot\HelpGeneration\Exceptions\NewValidateHelpExceptions.csv"
-
-    $Exceptions | ft
+    
+    if (($Exceptions | Measure-Object).Count -gt 0)
+    {
+        $Exceptions | ft
+        throw "A markdown file containing the help for a cmdlet is incomplete. Please check the exceptions provided for more details."
+    }
 
     Remove-Item -Path "$PSScriptRoot\HelpGeneration\Exceptions\NewValidateHelpExceptions.csv" -Force
 }
